Started Lab_4.py, passed Lab_7.py, renamed all them by status

+++ Lab_6(half-done)/Lab_6.py
@@ -41,7 +41,7 @@ class Sort():
 
 
 sort = Sort()
-array = [int(i)
+array = [float(i)
          for i in input("Введіть елементи масиву через пробіл:").split()]
 print("Ваш масив:", *array)
 result1 = sort.QuickSort(array)
